+++ src/main/java/com/solstice/stocks/service/StockServices.java
@@ -1,4 +1,4 @@
-package com.solstice.stocks;
+package com.solstice.stocks.service;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
